# make resolver context non-pointer