--- a/scalars.go
+++ b/scalars.go
@@ -51,7 +51,7 @@ var DateTimeType = &graphql.ScalarType{
 func NonZeroDateTime(fieldName string) *graphql.FieldDefinition {
 	return &graphql.FieldDefinition{
 		Type: DateTimeType,
-		Resolve: func(ctx *graphql.FieldContext) (interface{}, error) {
+		Resolve: func(ctx graphql.FieldContext) (interface{}, error) {
 			if t := fieldValue(ctx.Object, fieldName).(time.Time); !t.IsZero() {
 				return t, nil
 			}