## What it Does

This is a breaking change to make resolvers take a `FieldContext` argument instead of a `*FieldContext`.

Just removing the pointer is a significant improvement in performance:

Before:

```
% go test -v ./graphql/executor -bench=BenchmarkExecuteRequest -run asdqweqwe -benchtime 5s
goos: darwin
goarch: arm64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest
BenchmarkExecuteRequest-10    	    5815	    876965 ns/op	  796542 B/op	   20444 allocs/op
```

After:

```
% go test -v ./graphql/executor -bench=BenchmarkExecuteRequest -run asdqweqwe -benchtime 5s
goos: darwin
goarch: arm64
pkg: github.com/ccbrown/api-fu/graphql/executor
BenchmarkExecuteRequest
BenchmarkExecuteRequest-10    	    7444	    806404 ns/op	  665852 B/op	   18402 allocs/op
```

That's a big improvement in allocations and almost a 10% improvement in speed.

## Steps to Test

<!-- All changes should have automated tests when feasible: -->

`go test -v ./...` and see above for benchmark command.

<!-- Does running this require any special setup or dependencies? -->